Removed unncessary arrays

+++ include/mft.h
@@ -12,7 +12,7 @@ typedef union {
   uint8_t raw[MFT_SIZE];
 } MFT;
 
-int load_mft(int, MFT*, VBR*, off_t);
-int load_mft_mirror(int, MFT*, VBR*, off_t);
+ssize_t load_mft(int, MFT*, VBR*, off_t);
+ssize_t load_mft_mirror(int, MFT*, VBR*, off_t);
 
 #endif
